Add Flask json encoder
